Split tasks command into tasks and groups commands
