+++ cloudnet/src/main/resources/lang/english.properties
@@ -80,7 +80,8 @@ command-description-clear=Clears the entire console to get a better overview
 #
 # Console Command description defaults
 #
-command-description-tasks=Administers the configurations of all persistent tasks and groups
+command-description-tasks=Administers the configurations of all persistent tasks
+command-description-groups=Administers the configurations of all persistent groups
 command-description-service=Manages all local services that are provided
 command-description-create=Creates one or more new services based on a task or completely independent
 command-description-cluster=Gives an overview of what is known in the cluster
@@ -104,6 +105,16 @@ command-screen-enable-for-service=The automatic logging of the console protocol 
 command-screen-disable-for-service=The automatic logging of the console protocol for the service %uniqueId%:%name% has been disabled.
 command-screen-write-success=You've successfully sent the command "%command%" to the services %targets%
 command-screen-write-no-screen=There is no enabled screen to send the command to
+# command tasks/groups
+command-service-base-add-template-success=The template was successfully found and was added to the configuration
+command-service-base-add-inclusion-success=The integration was successfully added to the configuration.
+command-service-base-add-deployment-success=The swap was successfully added to the configuration
+command-service-base-add-group-no-group-found=No group with that name was found
+command-service-base-remove-template-success=The template was successfully removed from the configuration
+command-service-base-remove-inclusion-success=The integration was successfully removed from the configuration.
+command-service-base-remove-deployment-success=The swap was successfully removed from the configuration
+command-service-base-group-not-found=That group doesn't exist!
+command-service-base-create-group=The empty group was successfully created. Configure it via the command 'tasks' or in the configuration file
 # command tasks
 command-tasks-setup-no-console=This command is only available for the console!
 command-tasks-setup-question-name=What's the name of the new task?
@@ -119,30 +130,20 @@ command-tasks-setup-question-application=Which ServiceVersion should be ran on s
 command-tasks-setup-question-nodes=On which nodes can the services of this task start? (separated by ";")
 command-tasks-setup-create-begin=Trying to create the new task...
 command-tasks-setup-create-success=The new task %name% has been successfully created!
-command-tasks-create-group=The empty group was successfully created. Configure it via the command 'tasks' or in the configuration file
-command-tasks-create-task=The empty task was successfully created. Configure it via the command 'tasks' or in the configuration file
-command-tasks-delete-group=The group was successfully deleted
-command-tasks-delete-task=The task was successfully deleted
-
-command-tasks-reload-success=The ServiceTask and Group configurations have been reloaded!
-command-tasks-task-not-found=That task doesn't exist!
-command-tasks-add-template-success=The template was successfully found and was added to the configuration
-command-tasks-add-template-create-failed=The template could not be created
-command-tasks-add-inclusion-success=The integration was successfully added to the configuration.
-command-tasks-add-deployment-success=The swap was successfully added to the configuration
 command-tasks-add-group-success=The group was successfully added to the task
-command-tasks-add-group-no-group-found=No group with that name was found
 command-tasks-add-node-success=The node was successfully added to the task
 command-tasks-node-not-found=That node doesn't exist!
-command-tasks-group-not-found=That group doesn't exist!
-command-tasks-task-already-existing=That task already exists!
-command-tasks-group-already-existing=That group already exists!
-command-tasks-remove-template-success=The template was successfully removed from the configuration
-command-tasks-remove-inclusion-success=The integration was successfully removed from the configuration.
-command-tasks-remove-deployment-success=The swap was successfully removed from the configuration
 command-tasks-remove-group-success=The group was successfully removed from the task
 command-tasks-remove-node-success=The node was successfully removed from the task. All new services of this task won't start on this node
+command-tasks-create-task=The empty task was successfully created. Configure it via the command 'tasks' or in the configuration file
+command-tasks-delete-task=The task was successfully deleted
+command-tasks-task-already-existing=That task already exists!
 command-tasks-set-property-success=You have changed the configuration %property% from the group %name% to %value%.
+command-tasks-reload-success=The ServiceTask and Group configurations have been reloaded!
+command-tasks-task-not-found=That task doesn't exist!
+# command groups
+command-groups-delete-group=The group was successfully deleted
+command-groups-group-already-existing=That group already exists!
 # command service
 command-service-service-not-found=That service doesn't exist
 command-service-add-template-success=The template was successfully found and was added to the waiting templates
