+++ cloudnet/src/main/resources/lang/german.properties
@@ -80,7 +80,8 @@ command-description-clear=Leert die gesamte Konsole um eine bessere Übersicht z
 #
 # Console Command description defaults
 #
-command-description-tasks=Verwaltet alle persistenten Tasks und Gruppen, dessen Konfigurationen
+command-description-tasks=Verwaltet alle persistenten Tasks
+command-description-groups=Verwaltet alle persistenten Gruppen
 command-description-service=Verwaltet alle lokalen Services
 command-description-create=Erstellt einen oder mehrere neue Services auf Basis eines Tasks oder komplett eigenständig
 command-description-cluster=Gibt eine Übersicht über alles im Cluster bekannte
@@ -104,6 +105,16 @@ command-screen-enable-for-service=Das automatische Logging des Konsolenprotokoll
 command-screen-disable-for-service=Das automatische Logging des Konsolenprotokolls für den Service %uniqueId%:%name% wurde deaktiviert.
 command-screen-write-success=Der Befehl "%command%" wurde erfolgreich an die Services %targets% gesendet
 command-screen-write-no-screen=Du bist in keinem Screen
+# command tasks/groups
+command-service-base-add-template-success=Die Vorlage wurde erfolgreich gefunden und wurde in der Konfiguration hinzugefügt
+command-service-base-add-inclusion-success=Die Einbindung wurde erfolgreich der Konfiguration hinzugefügt
+command-service-base-add-deployment-success=Die Auslagerung wurde erfolgreich in die Konfiguration hinzugefügt
+command-service-base-add-group-no-group-found=Diese Gruppe existiert nicht
+command-service-base-remove-template-success=Die Vorlage wurde erfolgreich von der der Konfiguration entfernt
+command-service-base-remove-inclusion-success=Die Einbindung wurde erfolgreich von der Konfiguration entfernt
+command-service-base-remove-deployment-success=Die Auslagerung wurde erfolgreich von der Konfiguration entfernt
+command-service-base-group-not-found=Diese Gruppe existiert nicht!
+command-service-base-create-group=Die leere Gruppe wurde erfolgreich erstellt. Konfigurieren sie diese, via dem Befehl 'tasks' oder in der Konfiugrationsdatei
 # command tasks
 command-tasks-setup-no-console=Dieser Befehl ist nur in der Konsole verfügbar!
 command-tasks-setup-task-already-exists=Dieser Task existiert bereits!
@@ -119,30 +130,20 @@ command-tasks-setup-question-application=Welche Service Version soll auf den Ser
 command-tasks-setup-question-nodes=Auf welchen Nodes können Services dieses Tasks gestartet werden? (Getrennt mit ";")
 command-tasks-setup-create-begin=Der Task wird erstellt...
 command-tasks-setup-create-success=Der neue Task %name% wurde erfolgreich erstellt!
-command-tasks-create-group=Die leere Gruppe wurde erfolgreich erstellt. Konfigurieren sie diese, via dem Befehl 'tasks' oder in der Konfiugrationsdatei
-command-tasks-create-task=Der leere Task wurde erfolgreich erstellt. Konfigurieren sie diese, via dem Befehl 'tasks' oder in der Konfigurationsdatei
-command-tasks-delete-group=Die Gruppe wurde erfolgreich gelöscht
-command-tasks-delete-task=Der Task wurde erfolgreich gelöscht
-
-command-tasks-reload-success=Die Tasks und Gruppen Konfigurationen wurden neu geladen!
-command-tasks-task-not-found=Dieser Task existiert nicht!
-command-tasks-add-template-success=Die Vorlage wurde erfolgreich gefunden und wurde in der Konfiguration hinzugefügt
-command-tasks-add-template-create-failed=Die Vorlage konnte nicht erstellt werden
-command-tasks-add-inclusion-success=Die Einbindung wurde erfolgreich der Konfiguration hinzugefügt
-command-tasks-add-deployment-success=Die Auslagerung wurde erfolgreich in die Konfiguration hinzugefügt
 command-tasks-add-group-success=Die Gruppe wurde dem Task erfolgreich hinzugefügt
-command-tasks-add-group-no-group-found=Diese Gruppe existiert nicht
 command-tasks-add-node-success=Dem Task wurde erfolgreich die Node zugewiesen
 command-tasks-node-not-found=Diese Node existiert nicht!
-command-tasks-group-not-found=Diese Gruppe existiert nicht!
-command-tasks-task-already-existing=Dieser Task existiert bereits!
-command-tasks-group-already-existing=Diese Gruppe existiert bereits!
-command-tasks-remove-template-success=Die Vorlage wurde erfolgreich von der der Konfiguration entfernt
-command-tasks-remove-inclusion-success=Die Einbindung wurde erfolgreich von der Konfiguration entfernt
-command-tasks-remove-deployment-success=Die Auslagerung wurde erfolgreich von der Konfiguration entfernt
 command-tasks-remove-group-success=Die Gruppe wurde erfolgreich von dem Task entfernt
-command-tasks-remove-node-success=Der Task wurde erfolgreich von der Node entfernt. Alle neuen Services, werden nicht auf dieser Node bereitgestellt.
+command-tasks-remove-node-success=Der Task wurde erfolgreich von der Node entfernt. Alle neuen Services werden nicht auf dieser Node bereitgestellt.
+command-tasks-create-task=Der leere Task wurde erfolgreich erstellt. Konfigurieren sie diese, via dem Befehl 'tasks' oder in der Konfigurationsdatei
+command-tasks-delete-task=Der Task wurde erfolgreich gelöscht
+command-tasks-task-already-existing=Dieser Task existiert bereits!
 command-tasks-set-property-success=Du hast die Konfiguration %property% von der Gruppe %name% auf %value% umgestellt
+command-tasks-reload-success=Die Tasks und Gruppen Konfigurationen wurden neu geladen!
+command-tasks-task-not-found=Dieser Task existiert nicht!
+# command groups
+command-groups-delete-group=Die Gruppe wurde erfolgreich gelöscht
+command-groups-group-already-existing=Diese Gruppe existiert bereits!
 # command service
 command-service-service-not-found=Dieser Service existiert nicht
 command-service-add-template-success=Die Vorlage wurde erfolgreich gefunden und wurde zu den wartenden Vorlagen hinzugefügt
